Renamed week_06 to prac_06
