+++ prac_06/car.py
