+++ prac_06/guitars.py
@@ -1,6 +1,6 @@
 """CP1404 Practical - Guitars prac"""
 
-from week_06.guitar import Guitar
+from prac_06.guitar import Guitar
 
 def main():
     guitars = []
