+++ prac_06/languages.py
@@ -1,6 +1,6 @@
 """CP1404 Practical - Languages prac"""
 
-from week_06.programming_language import ProgrammingLanguage
+from prac_06.programming_language import ProgrammingLanguage
 
 def main():
