+++ prac_06/programming_language.py
