+++ prac_06/used_cars.py
@@ -1,7 +1,7 @@
 """CP1404/CP5632 Practical - Client code to use the Car class."""
 # Note that the import has a folder (module) in it.
 
-from week_06.car import Car
+from prac_06.car import Car
 
 
 def main():
